% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{crossval}
\alias{crossval}
\title{Title}
\usage{
crossval(A, nb.boot = 5, thresh = 1e-05, maxit = 100,
  lambda1.max = NULL, lambda2.max = NULL, lambda1.min = NULL,
  lambda2.min = NULL, length = 10, S0 = NULL, L0 = NULL,
  trace.it = T)
}
\arguments{
\item{A}{nxp heterogeneous observations (matrix)}

\item{nb.boot}{number of folds for cross validation (integer)}

\item{thresh}{convergence tolerance (positive number)}

\item{maxit}{maximum number of iterations (positive integer)}

\item{lambda1.max}{maximum regularization parameter for nuclear norm penalty (positive number)}

\item{lambda2.max}{maximum regularization parameter for ell_1 norm penalty (positive number)}

\item{lambda1.min}{minimum regularization parameter for nuclear norm penalty (positive number)}

\item{lambda2.min}{minimum regularization parameter for ell_1 norm penalty (positive number)}

\item{length}{size of cross-validation grid (integer)}

\item{trace.it}{whether messages about convergence should be printed (boolean)}

\item{groups}{vector of size n of group memberships (factor)}

\item{var.type}{vector of size p indicating types of columns (currently "gaussian", "binomial" or "poisson")}

\item{alpha0}{initial value for the main effects (matrix of size (number of groups)xp)}

\item{theta0}{initial value for the interactions (matrix of size nxp)}
}
\description{
Title
}
