% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsbm_mcgd.R
\name{gsbm_mcgd}
\alias{gsbm_mcgd}
\title{Fit sparse main effects plus low-rank interactions model
with least-squares data fitting term (numeric variables only)}
\usage{
gsbm_mcgd(A, lambda1, lambda2, epsilon, U = NULL, maxit = 100,
  thresh = 1e-06, S0 = NULL, L0 = NULL, R0 = NULL, trace.it = T)
}
\arguments{
\item{A}{nxn adjacency matrix}

\item{lambda1}{regularization parameter for nuclear norm penalty (positive number)}

\item{lambda2}{regularization parameter for ell_2,1 penalty (positive number)}

\item{U}{lower bound on nuclear norm (positive number, if NULL, default method is applied)}

\item{maxit}{maximum number of iterations (positive integer)}

\item{thresh}{convergence tolerance (positive number)}

\item{S0}{initial value for the sparse component}

\item{L0}{initial value for the low-rank component}

\item{R0}{lower bound on nuclear norm of L0 (positive number, if NULL, default method is applied)}

\item{trace.it}{whether messages about convergence should be printed (boolean)}
}
\description{
Fit sparse main effects plus low-rank interactions model
with least-squares data fitting term (numeric variables only)
}
